requirements.txt
Flask>=2.0.0
google-generativeai>=0.0.0
python-dotenv>=0.10.0
openai>=1.78.0
websockets>=15.0.1
requests>=2.32.3 